tools/agents/requirements.txt
aider-chat==0.69.1
# opencode is not distributed on PyPI; install separately per docs/agentic-harness.md
